New component modal - Used old mobile nav menu as basis for new component
