+++ src/layout/modal/Modal.module.css
@@ -1,0 +1,26 @@
+.wrapper {
+    display: block;
+}
+
+.active {
+    display: block;
+}
+
+
+.container {
+        width: 100%;
+        height: 100%;
+        position: absolute;
+        left: 0;
+        top: 0;
+        background-color: var(--background-secondary);
+        transform: translateX(0%);
+        transition: transform 300ms ease-in-out;
+        z-index: 98;
+        padding: 1.5rem;
+}
+
+.container_hidden {
+        transform: translateX(100%);
+        transition: transform 300ms ease-in-out;
+}
